parser_app/requirements.txt
# requirements.txt
selenium>=4.0
pandas>=1.3
openpyxl>=3.0
pillow>=9.0
imagehash>=4.2
requests>=2.26